Fix bugs Cost Calculation from user feedback

+++ Com.Bateeq.Service.Merchandiser.Lib/Services/CostCalculationGarmentService.cs
@@ -14,6 +14,7 @@ using Com.Bateeq.Service.Merchandiser.Lib.ViewModels;
 using Com.Moonlay.NetCore.Lib.Service;
 using System.Reflection;
 using Com.Bateeq.Service.Merchandiser.Lib.Services.AzureStorage;
+using Com.Bateeq.Service.Merchandiser.Lib.Exceptions;
 
 namespace Com.Bateeq.Service.Merchandiser.Lib.Services
 {
@@ -138,20 +139,27 @@ namespace Com.Bateeq.Service.Merchandiser.Lib.Services
 
         public override async Task<int> DeleteModel(int Id)
         {
-            CostCalculationGarment_MaterialService CostCalculationGarment_MaterialService = this.ServiceProvider.GetService<CostCalculationGarment_MaterialService>();
-
-            HashSet<int> CostCalculationGarment_Materials = new HashSet<int>(CostCalculationGarment_MaterialService.DbSet
-                .Where(p => p.CostCalculationGarmentId.Equals(Id))
-                .Select(p => p.Id));
-
-            foreach (int CostCalculationGarment_Material in CostCalculationGarment_Materials)
-            {
-                await CostCalculationGarment_MaterialService.DeleteModel(CostCalculationGarment_Material);
-            }
-
             CostCalculationGarment deleted = await this.GetAsync(Id);
-            await this.AzureImageService.RemoveImage(deleted.GetType().Name, deleted.ImagePath);
+            if (deleted.RO_GarmentId != null)
+            {
+                throw new DbReferenceNotNullException("Cost Calculation Garment ini tidak bisa di delete karena masih terdaftar di RO");
+            }
+            else
+            {
+                CostCalculationGarment_MaterialService CostCalculationGarment_MaterialService = this.ServiceProvider.GetService<CostCalculationGarment_MaterialService>();
 
+                HashSet<int> CostCalculationGarment_Materials = new HashSet<int>(CostCalculationGarment_MaterialService.DbSet
+                    .Where(p => p.CostCalculationGarmentId.Equals(Id))
+                    .Select(p => p.Id));
+
+                foreach (int CostCalculationGarment_Material in CostCalculationGarment_Materials)
+                {
+                    await CostCalculationGarment_MaterialService.DeleteModel(CostCalculationGarment_Material);
+                }
+                
+                await this.AzureImageService.RemoveImage(deleted.GetType().Name, deleted.ImagePath);
+            }
+            
             return await this.DeleteAsync(Id);
         }
 
@@ -199,17 +207,17 @@ namespace Com.Bateeq.Service.Merchandiser.Lib.Services
             viewModel.Efficiency.Id = model.EfficiencyId;
             viewModel.Efficiency.Value = PercentageConverter.ToPercent(model.EfficiencyValue);
 
-            viewModel.Wage = new OTLViewModel();
+            viewModel.Wage = new RateViewModel();
             viewModel.Wage.Id = model.WageId;
-            viewModel.Wage.Rate = model.WageRate;
+            viewModel.Wage.Value = model.WageRate;
 
-            viewModel.THR = new OTLViewModel();
+            viewModel.THR = new RateViewModel();
             viewModel.THR.Id = model.THRId;
-            viewModel.THR.Rate = model.THRRate;
+            viewModel.THR.Value = model.THRRate;
 
-            viewModel.RateDollar = new OTLViewModel();
-            viewModel.RateDollar.Id = model.RateDollarId;
-            viewModel.RateDollar.Rate = model.RateDollarRate;
+            viewModel.Rate = new RateViewModel();
+            viewModel.Rate.Id = model.RateId;
+            viewModel.Rate.Value = model.RateValue;
 
             viewModel.CostCalculationGarment_Materials = new List<CostCalculationGarment_MaterialViewModel>();
             if (model.CostCalculationGarment_Materials != null)
@@ -265,15 +273,15 @@ namespace Com.Bateeq.Service.Merchandiser.Lib.Services
             viewModel.CommissionPortion = PercentageConverter.ToPercent(model.CommissionPortion);
             viewModel.Risk = PercentageConverter.ToPercent(model.Risk);
 
-            viewModel.OTL1 = new OTLCalculatedViewModel();
+            viewModel.OTL1 = new RateCalculatedViewModel();
             viewModel.OTL1.Id = model.OTL1Id;
-            viewModel.OTL1.Rate = model.OTL1Rate;
-            viewModel.OTL1.CalculatedRate = model.OTL1CalculatedRate;
+            viewModel.OTL1.Value = model.OTL1Rate;
+            viewModel.OTL1.CalculatedValue = model.OTL1CalculatedRate;
 
-            viewModel.OTL2 = new OTLCalculatedViewModel();
+            viewModel.OTL2 = new RateCalculatedViewModel();
             viewModel.OTL2.Id = model.OTL2Id;
-            viewModel.OTL2.Rate = model.OTL2Rate;
-            viewModel.OTL2.CalculatedRate = model.OTL2CalculatedRate;
+            viewModel.OTL2.Value = model.OTL2Rate;
+            viewModel.OTL2.CalculatedValue = model.OTL2CalculatedRate;
 
             viewModel.NETFOBP = PercentageConverter.ToPercent((double)model.NETFOBP);
 
@@ -302,13 +310,13 @@ namespace Com.Bateeq.Service.Merchandiser.Lib.Services
             model.EfficiencyValue = PercentageConverter.ToFraction(viewModel.Efficiency.Value);
 
             model.WageId = viewModel.Wage.Id;
-            model.WageRate = viewModel.Wage.Rate != null ? (double)viewModel.Wage.Rate : 0;
+            model.WageRate = viewModel.Wage.Value != null ? (double)viewModel.Wage.Value : 0;
 
             model.THRId = viewModel.THR.Id;
-            model.THRRate = viewModel.THR.Rate != null ? (double)viewModel.THR.Rate : 0;
+            model.THRRate = viewModel.THR.Value != null ? (double)viewModel.THR.Value : 0;
 
-            model.RateDollarId = viewModel.RateDollar.Id;
-            model.RateDollarRate = viewModel.RateDollar.Rate != null ? (double)viewModel.RateDollar.Rate : 0;
+            model.RateId = viewModel.Rate.Id;
+            model.RateValue = viewModel.Rate.Value != null ? (double)viewModel.Rate.Value : 0;
 
             model.CostCalculationGarment_Materials = new List<CostCalculationGarment_Material>();
 
@@ -334,12 +342,12 @@ namespace Com.Bateeq.Service.Merchandiser.Lib.Services
             model.Risk = PercentageConverter.ToFraction(viewModel.Risk);
 
             model.OTL1Id = viewModel.OTL1.Id;
-            model.OTL1Rate = viewModel.OTL1.Rate != null ? (double)viewModel.OTL1.Rate : 0;
-            model.OTL1CalculatedRate = viewModel.OTL1.CalculatedRate != null ? (double)viewModel.OTL1.CalculatedRate : 0;
+            model.OTL1Rate = viewModel.OTL1.Value != null ? (double)viewModel.OTL1.Value : 0;
+            model.OTL1CalculatedRate = viewModel.OTL1.CalculatedValue != null ? (double)viewModel.OTL1.CalculatedValue : 0;
 
             model.OTL2Id = viewModel.OTL2.Id;
-            model.OTL2Rate = viewModel.OTL2.Rate != null ? (double)viewModel.OTL2.Rate : 0;
-            model.OTL2CalculatedRate = viewModel.OTL2.CalculatedRate != null ? (double)viewModel.OTL2.CalculatedRate : 0;
+            model.OTL2Rate = viewModel.OTL2.Value != null ? (double)viewModel.OTL2.Value : 0;
+            model.OTL2CalculatedRate = viewModel.OTL2.CalculatedValue != null ? (double)viewModel.OTL2.CalculatedValue : 0;
 
             model.NETFOBP = PercentageConverter.ToFraction(viewModel.NETFOBP);
